[Sprint 1][Task 6] chore: update integration tests module structure
