+++ auth-service/tests/api/root.rs
@@ -1,0 +1,10 @@
+use crate::helpers::TestApp;
+
+// Tokio's test macro is used to run the test in an async environment
+#[tokio::test]
+async fn root_returns_auth_ui() {
+    let app = TestApp::new().await;
+    let response = app.get_root().await;
+
+    assert_eq!(response.status().as_u16(), 200);
+}
