+++ auth-service/tests/api/signup.rs
@@ -1,0 +1,10 @@
+use crate::helpers::TestApp;
+
+// Tokio's test macro is used to run the test in an async environment
+#[tokio::test]
+async fn signup_returns_auth_ui() {
+    let app = TestApp::new().await;
+    let response = app.post_signup().await;
+
+    assert_eq!(response.status().as_u16(), 200);
+}
