+++ auth-service/tests/api/verify_2fa.rs
@@ -1,0 +1,10 @@
+use crate::helpers::TestApp;
+
+// Tokio's test macro is used to run the test in an async environment
+#[tokio::test]
+async fn verify_2fa_returns_auth_ui() {
+    let app = TestApp::new().await;
+    let response = app.post_verify_2fa().await;
+
+    assert_eq!(response.status().as_u16(), 200);
+}
